Build a release asset for FreeBSD
This uses the `vmactions/freebsd-vm` GitHub Action to build an asset for
FreeBSD alongside the existing assets. See issue 438 for context.

The build steps as written produce a working binary on FreeBSD 14.0
x86_64. Note however that the VM used for CI is FreeBSD 13.2; this is
because binaries built on a FreeBSD system are forward- but not
backward-compatible with other FreeBSD versions. That is, an executable
built on 14.0 wouldn't necessarily work on 13.2, but one built on 13.2
will work on 14.0. The choice of 13.2 in particular is because it's the
oldest currently supported release.

Python 3.11 is used for consistency with other platforms. Unlike other
platforms, this build utilizes a virtual environment, simply because
it's easier and less error prone.